handle ACCEPTOR-SHUTDOWN-P correctly
After stopping an acceptor with STOP, one can't start it again with
START, because the START method doesn't reset the value of
acceptor-shutdown-p to NIL, so the listener returns/quits instantly.

Also, the default value of ACCEPTOR-SHUTDOWN-P should be T, because
every acceptor should be closed when it is created.

See [1] & [2] for details.

[1] http://article.gmane.org/gmane.lisp.lib.tbnl.general/2469
[2] http://article.gmane.org/gmane.lisp.lib.tbnl.general/2868